Adding Closures for Embedding and Parameter Optimization
