+++ popari/_parameter_optimizer.py
@@ -3,6 +3,7 @@ import numpy as np
 import torch
 from tqdm.auto import tqdm, trange
 
+from popari._parameter_optimizer_closures import *
 from popari._popari_dataset import PopariDataset
 from popari.sample_for_integral import integrate_of_exponential_over_simplex
 from popari.util import (
@@ -616,29 +617,18 @@ class ParameterOptimizer:
             for group_M_bar in M_bar:
                 differential_regularization_linear_term += group_weighting * self.lambda_M * group_M_bar
 
-        def compute_loss(M):
-            quadratic_factor_grad = M @ (quadratic_factor + differential_regularization_quadratic_factor)
-            loss = (quadratic_factor_grad * M).sum()
-            linear_term_grad = linear_term + differential_regularization_linear_term
-            loss -= 2 * (linear_term_grad * M).sum()
-
-            loss += constant
-
-            if self.metagene_mode == "differential" and M_bar is not None:
-                differential_regularization_term = (M @ differential_regularization_quadratic_factor * M).sum() - 2 * (
-                    differential_regularization_linear_term * M
-                ).sum()
-                group_weighting = 1 / len(M_bar)
-                for group_M_bar in M_bar:
-                    differential_regularization_term += (
-                        group_weighting * self.lambda_M * (group_M_bar * group_M_bar).sum()
-                    )
-
-            loss /= 2
-
-            return loss.item()
-
-        loss = compute_loss(M)
+        compute_loss = compute_loss_nll_M_closure(
+            M,
+            quadratic_factor,
+            differential_regularization_quadratic_factor,
+            linear_term,
+            differential_regularization_linear_term,
+            constant,
+            self.metagene_mode,
+            M_bar,
+            self.lambda_M,
+        )
+        loss = compute_loss
 
         return loss
 
@@ -739,95 +729,26 @@ class ParameterOptimizer:
         verbose_bar = tqdm(disable=not (self.verbose > 2), bar_format="{desc}{postfix}")
         progress_bar = trange(n_epochs, leave=True, disable=not self.verbose, desc="Updating M", miniters=1000)
 
-        def compute_loss_and_gradient(M):
-            quadratic_factor_grad = M @ (quadratic_factor + differential_regularization_quadratic_factor)
-            loss = (quadratic_factor_grad * M).sum()
-            verbose_description = ""
-            if self.verbose > 2:
-                verbose_description += f"M quadratic term: {loss:.1e}"
-            linear_term_grad = linear_factor + differential_regularization_linear_factor
-            loss -= 2 * (linear_term_grad * M).sum()
-            grad = quadratic_factor_grad - linear_term_grad
+        # compute_loss_and_gradient = compute_loss_and_gradient_M_closure(M, quadratic_factor, differential_regularization_quadratic_factor, self.verbose, differential_regularization_linear_factor, constant, self.metagene_mode, M_bar, self.lambda_M, self.M_constraint)
 
-            loss += constant
-
-            if self.metagene_mode == "differential" and M_bar is not None:
-                differential_regularization_term = (M @ differential_regularization_quadratic_factor * M).sum() - 2 * (
-                    differential_regularization_linear_factor * M
-                ).sum()
-                group_weighting = 1 / len(M_bar)
-                for group_M_bar in M_bar:
-                    differential_regularization_term += (
-                        group_weighting * self.lambda_M * (group_M_bar * group_M_bar).sum()
-                    )
-
-            if self.verbose > 2:
-                # print(f"M regularization term: {regularization_term}")
-                if self.metagene_mode == "differential":
-                    verbose_description += f"M differential regularization term: {differential_regularization_term}"
-
-            loss /= 2
-
-            if self.M_constraint == "simplex":
-                grad.sub_(grad.sum(0, keepdim=True))
-
-            return loss.item(), grad
-
-        def estimate_M_nag(M):
-            """Estimate M using Nesterov accelerated gradient descent.
-
-            Args:
-                M (torch.Tensor) : current estimate of meteagene parameters
-
-            """
-            loss, grad = compute_loss_and_gradient(M)
-            if self.verbose > 1:
-                print(f"M NAG Initial Loss: {loss}")
-
-            step_size = 1 / torch.linalg.eigvalsh(quadratic_factor).max().item()
-            loss = np.inf
-
-            optimizer = NesterovGD(M.clone(), step_size)
-            for epoch in progress_bar:
-                loss_prev = loss
-                M_prev = M.clone()
-
-                # Update M
-                loss, grad = compute_loss_and_gradient(M)
-                M = optimizer.step(grad)
-                if simplex_projection_mode == "exact":
-                    if self.use_inplace_ops:
-                        M = project_M_(M, self.M_constraint)
-                    else:
-                        M = project_M(M, self.M_constraint)
-                elif simplex_projection_mode == "approximate":
-                    raise NotImplementedError()
-
-                optimizer.set_parameters(M)
-
-                dloss = loss_prev - loss
-                dM = (M_prev - M).abs().max().item()
-                stop_criterion = dM < tol and epoch > 5
-                assert not np.isnan(loss)
-                if epoch % 5 == 0 or stop_criterion:
-                    description = (
-                        f"Updating M: loss = {loss:.1e}, "
-                        f"%δloss = {dloss / loss:.1e}, "
-                        f"δM = {dM:.1e}"
-                        # f'lr={step_size_scale:.1e}'
-                    )
-                    progress_bar.set_description(description)
-                if stop_criterion:
-                    break
-
-            verbose_bar.close()
-            progress_bar.close()
-
-            loss, grad = compute_loss_and_gradient(M)
-            if self.verbose > 1:
-                print(f"M NAG Final Loss: {loss}")
-
-            return M
+        estimate_M_nag = estimate_M_nag_closure(
+            M,
+            self.verbose,
+            quadratic_factor,
+            differential_regularization_quadratic_factor,
+            linear_factor,
+            differential_regularization_linear_factor,
+            constant,
+            self.metagene_mode,
+            M_bar,
+            self.lambda_M,
+            progress_bar,
+            simplex_projection_mode,
+            self.use_inplace_ops,
+            self.M_constraint,
+            tol,
+            verbose_bar,
+        )
 
         if backend_algorithm == "mu":
             for epoch in progress_bar:
@@ -862,7 +783,19 @@ class ParameterOptimizer:
         elif backend_algorithm == "gd":
             step_size = 1 / torch.linalg.eigvalsh(quadratic_factor).max().item()
             step_size_scale = 1
-            loss, grad = compute_loss_and_gradient(M)
+            loss, grad = compute_loss_and_gradient(
+                M,
+                quadratic_factor,
+                differential_regularization_quadratic_factor,
+                self.verbose,
+                linear_factor,
+                differential_regularization_linear_factor,
+                constant,
+                self.metagene_mode,
+                M_bar,
+                self.lambda_M,
+                self.M_constraint,
+            )
             dM = dloss = np.inf
             for epoch in progress_bar:
                 M_new = M.sub(grad, alpha=step_size * step_size_scale)
@@ -873,7 +806,19 @@ class ParameterOptimizer:
                         M = project_M(M_new, self.M_constraint)
                 elif simplex_projection_mode == "approximate":
                     pass
-                loss_new, grad_new = compute_loss_and_gradient(M_new)
+                loss_new, grad_new = compute_loss_and_gradient(
+                    M_new,
+                    quadratic_factor,
+                    differential_regularization_quadratic_factor,
+                    self.verbose,
+                    linear_factor,
+                    differential_regularization_linear_factor,
+                    constant,
+                    self.metagene_mode,
+                    M_bar,
+                    self.lambda_M,
+                    self.M_constraint,
+                )
                 if loss_new < loss or step_size_scale == 1:
                     dM = (M_new - M).abs().max().item()
                     dloss = loss - loss_new
@@ -897,7 +842,7 @@ class ParameterOptimizer:
                     break
 
         elif backend_algorithm == "gd Nesterov":
-            M = estimate_M_nag(M)
+            M = estimate_M_nag
         else:
             raise NotImplementedError
